refactor(graph): extract decay+prune into graph_memory/pruning.py
graph.py was 1522 lines with no internal module boundaries, making it
hard to find and safely change any one piece of logic. Splitting it
into cohesive modules (matching the existing types.py/helpers.py/
visualization.py pattern) so future changes touch a smaller, focused
file instead of the whole class.

Pure code motion: apply_importance_decay() and prune() move to
pruning.py as module-level functions taking the GraphMemory instance
as their first argument; graph.py keeps one-line delegating wrappers.
No logic changes — full suite (437 tests) passes unchanged.
